requirements.txt
# ProjectionAI - Python Dependencies

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Machine Learning
xgboost==2.0.2
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
eventlet==0.33.3

# Data Visualization
matplotlib==3.8.2
plotly==5.18.0

# HTTP Requests
requests==2.31.0
beautifulsoup4==4.12.2

# Web Scraping
selenium==4.16.0
webdriver-manager==4.0.1

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
python-json-logger==2.0.7

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3.post1

# Jupyter (for notebooks)
jupyter==1.0.0
ipython==8.18.1
ipykernel==6.28.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
